feat: share file context menu across preview and transcript
